chore: move database files to lib/server
