+++ package.json
@@ -15,12 +15,15 @@
 		"db:seed": "prisma db seed"
 	},
 	"prisma": {
-		"seed": "node prisma/seed.js"
+		"schema": "src/lib/server/database/prisma/schema.prisma",
+		"seed": "node src/lib/server/database/prisma/seed.js"
 	},
 	"devDependencies": {
 		"@sveltejs/adapter-node": "^5.0.1",
 		"@sveltejs/kit": "^2.5.10",
 		"@sveltejs/vite-plugin-svelte": "^3.1.1",
+		"@types/node": "^22.5.5",
+		"@types/nodemailer": "^6.4.15",
 		"sass": "^1.72.0",
 		"svelte": "^4.2.17",
 		"svelte-preprocess": "^5.1.3",
@@ -30,8 +33,6 @@
 	},
 	"dependencies": {
 		"@prisma/client": "^5.11.0",
-		"@types/node": "^22.5.5",
-		"@types/nodemailer": "^6.4.15",
 		"chalk": "^5.3.0",
 		"dotenv": "^16.4.5",
 		"nodemailer": "^6.9.13",
