+++ src/lib/server/database/prisma/schema.prisma
@@ -8,7 +8,7 @@ generator client {
 // Define database provider and url
 datasource db {
     provider = "sqlite"
-    url      = "file:./database.sqlite"
+    url      = "file:../database.sqlite"
 }
