+++ src/routes/(main)/(private)/(user)/settings/+page.server.js
@@ -1,5 +1,5 @@
 // #region Imports
-import dbClient from "$lib/server/prisma.js"
+import dbClient from "$lib/server/database/prisma/prisma.js"
 import inputHandler from "$lib/server/inputHandler.js"
 import { stringHasher } from "$lib/server/hashUtils.js"
 import { emailer } from "$lib/server/emailUtils.js"
